[FIX] #21 - 뷰 구조 변경 mission을 하나의 secion에서 cell로 구조 변경
