+++ NotToDo/NotToDo/Presentation/HomeScene/Cells/HomeMissionCollectionViewCell.swift
@@ -2,7 +2,7 @@
 //  HomeMissionCollectionViewCell.swift
 //  NotToDo
 //
-//  Created by 강윤서 on 2023/01/07.
+//  Created by 강윤서 on 2023/01/09.
 //
 
 import UIKit
@@ -11,89 +11,187 @@ import SnapKit
 import Then
 
 final class HomeMissionCollectionViewCell: UICollectionViewCell {
-
-    // MARK: - Properties
-
+    
+    // MARK: Properties
+    
     static let identifier = "HomeMissionCollectionViewCell"
-
+    
     // MARK: - UI Components
-
-    private let backView = UIView()
-    private let yellowView = UIView()
-    private let behaviorLabel = UILabel()
-
-    // MARK: - Life Cycle
-
+    
+    private let headerBackView = UIView()
+    private let headerView = UIView()
+    private var statusButton = UIButton()
+    private let situationLabel = UILabel()
+    private let missionTitleLabel = UILabel()
+    private let meatballButton = UIButton()
+    private let goalView = UIView()
+    private let goalTitleLabel = UILabel()
+    private let goalLabel = UILabel()
+    
+    private let solutionStackview = UIStackView()
+    private var firstSolusionView = UIImageView()
+    private var secondSolustionView = UIImageView()
+    private var firstSolutionLabel = UILabel()
+    private var secondSolutionLabel = UILabel()
+    
+    // MARK: Life Cycle
+    
     override init(frame: CGRect) {
         super.init(frame: frame)
         setUI()
         setLayout()
-        prepare()
     }
-
+    
     @available(*, unavailable)
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
-    }
-
-    override func prepareForReuse() {
-        super.prepareForReuse()
-        prepare()
     }
 }
 
 extension HomeMissionCollectionViewCell {
     private func setUI() {
         backgroundColor = .BG
-        setBorder()
-        backView.backgroundColor = .nottodoWhite
-        yellowView.backgroundColor = .yellow_basic
-
-        behaviorLabel.do {
+        headerBackView.backgroundColor = .nottodoGray2
+        
+        headerView.do {
+            $0.backgroundColor = .nottodoWhite
+        }
+        
+        situationLabel.do {
+            $0.font = .PretendardRegular(size: 16.adjusted)
             $0.textColor = .nottodoGray1
-            $0.font = .PretendardRegular(size: 14.adjusted)
+        }
+        
+        missionTitleLabel.do {
+            $0.font = .PretendardBold(size: 16.adjusted)
+            $0.textColor = .nottodoBlack
+        }
+        
+        goalView.do {
+            $0.backgroundColor = .nottodoBlack
+            $0.layer.cornerRadius = 10
+        }
+        
+        goalLabel.do {
+            $0.textColor = .yellow_basic
+            $0.text = I18N.goal
+            $0.font = .PretendardRegular(size: 12.adjusted)
+        }
+        
+        goalTitleLabel.do {
+            $0.textColor = .nottodoGray1
+            $0.font = .PretendardMedium(size: 12.adjusted)
+        }
+        
+        meatballButton.setBackgroundImage(.dot, for: .normal)
+        statusButton.setBackgroundImage(.checkDefault, for: .normal)
+        
+        solutionStackview.do {
+            $0.distribution = .fill
+            $0.axis = .vertical
+            $0.spacing = 0
+        }
+        
+        firstSolusionView.image = .firstSolutionView
+        secondSolustionView.image = .secondSolutionView
+        
+        [firstSolutionLabel, secondSolutionLabel].forEach {
+            $0.textColor = .nottodoGray1
+            $0.font = .PretendardBold(size: 14.adjusted)
         }
     }
-
+    
     private func setLayout() {
-        backView.addSubviews(yellowView, behaviorLabel)
-        contentView.addSubview(backView)
+        headerBackView.addSubview(headerView)
+        headerView.addSubviews(statusButton, situationLabel, missionTitleLabel,
+                             goalView, goalTitleLabel, meatballButton)
+        goalView.addSubview(goalLabel)
+        
+        firstSolusionView.addSubview(firstSolutionLabel)
+        secondSolustionView.addSubview(secondSolutionLabel)
+        solutionStackview.addArrangedSubviews(firstSolusionView, secondSolustionView)
+        addSubviews(headerBackView, solutionStackview)
 
-        backView.snp.makeConstraints {
-            $0.trailing.equalToSuperview().inset(20.adjusted)
-            $0.leading.equalToSuperview().inset(51.adjusted)
-            $0.top.bottom.trailing.equalToSuperview()
+        headerBackView.snp.makeConstraints {
+            $0.top.equalToSuperview().inset(17.adjusted)
+            $0.height.equalTo(81.adjusted)
+            $0.leading.trailing.equalToSuperview().inset(20.adjusted)
         }
-
-        yellowView.snp.makeConstraints {
-            $0.top.bottom.leading.equalToSuperview()
-            $0.width.equalTo(11.adjusted)
+        
+        headerView.snp.makeConstraints {
+            $0.top.bottom.leading.trailing.equalToSuperview().inset(0.5.adjusted)
         }
-
-        behaviorLabel.snp.makeConstraints {
+        
+        goalLabel.snp.makeConstraints {
+            $0.center.equalToSuperview()
+        }
+        
+        statusButton.snp.makeConstraints {
+            $0.leading.equalToSuperview().inset(19.adjusted)
             $0.centerY.equalToSuperview()
-            $0.leading.equalTo(yellowView.snp.trailing).offset(11.adjusted)
+        }
+        
+        situationLabel.snp.makeConstraints {
+            $0.top.equalToSuperview().inset(18.adjusted)
+            $0.leading.equalTo(statusButton.snp.trailing).offset(15.adjusted)
+        }
+        
+        missionTitleLabel.snp.makeConstraints {
+            $0.top.equalTo(situationLabel.snp.top)
+            $0.leading.equalTo(situationLabel.snp.trailing).offset(5.adjusted)
+        }
+        
+        goalView.snp.makeConstraints {
+            $0.top.equalTo(missionTitleLabel.snp.bottom).offset(8.adjusted)
+            $0.leading.equalTo(situationLabel.snp.leading)
+            $0.height.equalTo(20.adjusted)
+            $0.width.equalTo(51.adjusted)
+        }
+        
+        goalTitleLabel.snp.makeConstraints {
+            $0.top.equalTo(missionTitleLabel.snp.bottom).offset(11.adjusted)
+            $0.leading.equalTo(goalView.snp.trailing).offset(5.adjusted)
+        }
+        
+        meatballButton.snp.makeConstraints {
+            $0.trailing.equalToSuperview().inset(25.adjusted)
+            $0.centerY.equalToSuperview()
+        }
+        
+        solutionStackview.snp.makeConstraints {
+            $0.top.equalTo(headerView.snp.bottom).offset(7.adjusted)
+            $0.leading.equalToSuperview().inset(51.adjusted)
+            $0.trailing.equalTo(headerBackView.snp.trailing)
+            $0.bottom.equalToSuperview()
+        }
+        
+        [firstSolusionView, secondSolustionView].forEach {
+            $0.snp.makeConstraints {
+                $0.height.equalTo(40.adjusted)
+            }
+        }
+        
+        [firstSolutionLabel, secondSolutionLabel].forEach {
+            $0.snp.makeConstraints {
+                $0.centerY.equalToSuperview()
+                $0.leading.equalToSuperview().inset(22.adjusted)
+            }
         }
     }
-
-    func prepare() {
-        setBorder()
+    
+    func configure(_ model: DailyMission) {
+        situationLabel.text = model.situation
+        missionTitleLabel.text = model.title
+        goalTitleLabel.text = model.goal
+        setAction(model.actions)
     }
-
-    func setBorder() {
-        print("추가됨")
-        backView.layer.addBorder([.top, .left, .right, .bottom], color: .nottodoGray2!, width: 0.5)
-        yellowView.layer.addBorder([.top, .right], color: .nottodoGray2!, width: 0.5)
-//        if index.row == 0 {
-//            print("함수실행")
-//
-//        } else {
-//            backView.layer.addBorder([.top, .left, .right, .bottom], color: .nottodoGray2!, width: 0.5)
-//            yellowView.layer.addBorder([.top, .left, .right, .bottom], color: .nottodoGray2!, width: 0.5)
-//        }
-    }
-
-    func configure(_ text: String) {
-        behaviorLabel.text = text
+    
+    private func setAction(_ actionList: [String]) {
+        secondSolustionView.isHidden = actionList.count != 2
+        if actionList.count >= 2 {
+            secondSolutionLabel.text = actionList[1]
+            secondSolustionView.image = .secondSolutionView
+        }
+        firstSolutionLabel.text = actionList[0]
     }
 }
